requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
datasets==4.8.2
decorator==5.2.1
dill==0.4.1
executing==2.2.1
fastapi==0.135.2
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
hyperframe==6.1.0
idna==3.11
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qdrant-client==1.17.1
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.23.0
